Normalement fin de vérif des get et debut de déplacement des requêtes dans fichier à part pour index

+++ php/supprimer.php
@@ -11,10 +11,12 @@ require_once 'connect.php'; ?>
     include 'templatesHTML/navbar.php';
     if (isset($_GET['id'])) {
       $_SESSION['id_hist'] = (int) htmlspecialchars($_GET['id'], ENT_QUOTES, 'UTF-8', false);
+    }
+    if (isset($_GET['nomHist'])) {
       $_SESSION['nom_hist'] = htmlspecialchars($_GET['nomHist'], ENT_QUOTES, 'UTF-8', false);
     }
     if ($BDD) {
-      if (isset($_POST['suppr'])) {
+      if (isset($_POST['suppr']) && isset($_SESSION['id_hist'])) {
         $req = $BDD->prepare("DELETE FROM page_hist WHERE id_hist=:numero");
         $req->execute(array(
           "numero" => $_SESSION['id_hist']
